Refactor SafeMath32 library to use uint32 for arithmetic operations

+++ contracts/SafeMath32.sol
@@ -10,11 +10,11 @@ library SafeMath32 {
   /**
   * @dev Multiplies two numbers, throws on overflow.
   */
-  function mul(uint256 a, uint256 b) internal pure returns (uint256) {
+  function mul(uint32 a, uint32 b) internal pure returns (uint32) {
     if (a == 0) {
       return 0;
     }
-    uint256 c = a * b;
+    uint32 c = a * b;
     assert(c / a == b);
     return c;
   }
@@ -22,9 +22,9 @@ library SafeMath32 {
   /**
   * @dev Integer division of two numbers, truncating the quotient.
   */
-  function div(uint256 a, uint256 b) internal pure returns (uint256) {
+  function div(uint32 a, uint32 b) internal pure returns (uint32) {
     // assert(b > 0); // Solidity automatically throws when dividing by 0
-    uint256 c = a / b;
+    uint32 c = a / b;
     // assert(a == b * c + a % b); // There is no case in which this doesn't hold
     return c;
   }
@@ -32,7 +32,7 @@ library SafeMath32 {
   /**
   * @dev Subtracts two numbers, throws on overflow (i.e. if subtrahend is greater than minuend).
   */
-  function sub(uint256 a, uint256 b) internal pure returns (uint256) {
+  function sub(uint32 a, uint32 b) internal pure returns (uint32) {
     assert(b <= a);
     return a - b;
   }
@@ -40,8 +40,8 @@ library SafeMath32 {
   /**
   * @dev Adds two numbers, throws on overflow.
   */
-  function add(uint256 a, uint256 b) internal pure returns (uint256) {
-    uint256 c = a + b;
+  function add(uint32 a, uint32 b) internal pure returns (uint32) {
+    uint32 c = a + b;
     assert(c >= a);
     return c;
   }
